0.1.7 feat: Add update check on app launch
This commit adds a new feature to check for updates on app launch. The `fetchLatestTag` function in the `VersionService` class is called, which fetches the latest tag from the GitHub API. If an update is available, the user will be presented with an alert.

The following files were modified:
- App.swift: Added code to fetch latest tag and handle update availability.
- versionService.swift: Updated currentTag value to "/v0.1.6".
- ContentView.swift: Removed unnecessary print statement.

Note: The GitHub token used in versionService.swift has been updated for security reasons.